Restored obsolete APIs `DisableTimeoutManager`  which are being removed
