+++ src/Tests/ApprovalFiles/APIApprovals.ApproveAzureStorageQueueTransport.approved.txt
@@ -92,6 +92,9 @@ namespace NServiceBus
             ". The member currently throws a NotImplementedException. Will be removed in vers" +
             "ion 12.0.0.", true)]
         public void DisableDelayedDelivery() { }
+        [System.Obsolete("The member currently throws a NotImplementedException. Will be removed in version" +
+            " 12.0.0.", true)]
+        public void DisableTimeoutManager() { }
         [System.Obsolete("Configure the transport via the AzureStorageQueueTransport DelayedDelivery.Delaye" +
             "dDeliveryTableName property. Will be treated as an error from version 12.0.0. Wi" +
             "ll be removed in version 13.0.0.", false)]
