+++ src/Transport/obsoletes-v11.cs
@@ -1,5 +1,6 @@
 ﻿namespace NServiceBus
 {
+    using System;
     using global::Azure.Storage.Queues;
     using Microsoft.Azure.Cosmos.Table;
 
@@ -18,5 +19,38 @@
             TreatAsErrorFromVersion = "12.0",
             RemoveInVersion = "13.0")]
         public AccountInfo AddAccount(string alias, string connectionString) => AddAccount(alias, new QueueServiceClient(connectionString), CloudStorageAccount.Parse(connectionString).CreateCloudTableClient());
+    }
+
+    /// <summary>Configures native delayed delivery.</summary>
+    public partial class DelayedDeliverySettings
+    {
+        /// <summary>
+        /// Disables the Timeout Manager for the endpoint. Before disabling ensure there all timeouts in the timeout store
+        /// have been processed or migrated.
+        /// </summary>
+        [ObsoleteEx(
+            TreatAsErrorFromVersion = "11.0",
+            RemoveInVersion = "12.0")]
+        public void DisableTimeoutManager()
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Disable delayed delivery.
+        /// <remarks>
+        /// Disabling delayed delivery reduces costs associated with polling Azure Storage service for delayed messages that need
+        /// to be dispatched.
+        /// Do not use this setting if your endpoint requires delayed messages, timeouts, or delayed retries.
+        /// </remarks>
+        /// </summary>
+        [ObsoleteEx(
+            Message = "Configure delayed delivery support via the AzureStorageQueueTransport constructor.",
+            TreatAsErrorFromVersion = "11.0",
+            RemoveInVersion = "12.0")]
+        public void DisableDelayedDelivery()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
